chore: 체스 말 패키지 분리
