+++ src/test/java/chess/domain/piece/KingTest.java
@@ -1,7 +1,10 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.Path;
+import chess.domain.Position;
+import chess.domain.TeamColor;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -9,14 +12,15 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class RookTest {
+class KingTest {
+
 
     @Test
-    @DisplayName("Rook 은 앞뒤양옆으로 원하는 만큼 이동 가능하다.")
+    @DisplayName("Knight 은 앞뒤양옆 1칸 이동 후 대각선으로 1칸 이동 가능하다.")
     void 이동_범위_확인() {
-        Rook rook = new Rook(TeamColor.WHITE);
+        King king = new King(TeamColor.WHITE);
 
-        List<Path> movablePaths = rook.findMovablePaths(new Position(1, 1));
+        List<Path> movablePaths = king.findMovablePaths(new Position(1, 1));
 
         int totalPositionCount = 0;
 
@@ -24,6 +28,7 @@ class RookTest {
             totalPositionCount += path.positions().size();
         }
 
-        assertThat(totalPositionCount).isEqualTo(14);
+        assertThat(totalPositionCount).isEqualTo(3);
     }
+
 }
