+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,7 +1,10 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.Path;
+import chess.domain.Position;
+import chess.domain.TeamColor;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -9,15 +12,14 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class KingTest {
-
+class KnightTest {
 
     @Test
     @DisplayName("Knight 은 앞뒤양옆 1칸 이동 후 대각선으로 1칸 이동 가능하다.")
     void 이동_범위_확인() {
-        King king = new King(TeamColor.WHITE);
+        Knight knight = new Knight(TeamColor.WHITE);
 
-        List<Path> movablePaths = king.findMovablePaths(new Position(1, 1));
+        List<Path> movablePaths = knight.findMovablePaths(new Position(2, 2));
 
         int totalPositionCount = 0;
 
@@ -25,7 +27,7 @@ class KingTest {
             totalPositionCount += path.positions().size();
         }
 
-        assertThat(totalPositionCount).isEqualTo(3);
+        assertThat(totalPositionCount).isEqualTo(4);
     }
 
 }
