+++ src/test/java/chess/domain/piece/RookTest.java
@@ -1,7 +1,10 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.Path;
+import chess.domain.Position;
+import chess.domain.TeamColor;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -9,14 +12,14 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class KnightTest {
+class RookTest {
 
     @Test
-    @DisplayName("Knight 은 앞뒤양옆 1칸 이동 후 대각선으로 1칸 이동 가능하다.")
+    @DisplayName("Rook 은 앞뒤양옆으로 원하는 만큼 이동 가능하다.")
     void 이동_범위_확인() {
-        Knight knight = new Knight(TeamColor.WHITE);
+        Rook rook = new Rook(TeamColor.WHITE);
 
-        List<Path> movablePaths = knight.findMovablePaths(new Position(2, 2));
+        List<Path> movablePaths = rook.findMovablePaths(new Position(1, 1));
 
         int totalPositionCount = 0;
 
@@ -24,7 +27,6 @@ class KnightTest {
             totalPositionCount += path.positions().size();
         }
 
-        assertThat(totalPositionCount).isEqualTo(4);
+        assertThat(totalPositionCount).isEqualTo(14);
     }
-
 }
